pyproject.toml
[project]
name = "example-webhook-handler-py"
version = "0.1.0"
description = "Stripe-style webhook with event_id-as-promise-id deduplication — same event processed at most once."
readme = "README.md"
requires-python = ">=3.12"
license = "Apache-2.0"
dependencies = [
    "fastapi>=0.115.0",
    "resonate-sdk>=0.6.3",
    "uvicorn>=0.30.0",
]
